Rename StudentClass to Student
